+++ scrollytelling/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @bsmnt/scrollytelling
 
+## 0.2.4
+
+### Patch Changes
+
+- 56586af: Default disabled value to false
+- abe6bb9: Tweak disabled
+
+## 0.2.4-next.1
+
+### Patch Changes
+
+- Tweak disabled
+
+## 0.2.4-next.0
+
+### Patch Changes
+
+- Default disabled value to false
+
 ## 0.2.3
 
 ### Patch Changes